fix: add blog google analytics
